Add a RestrictedTransformer class
This behaves similarly to AffineTransformer, but takes more directly comprehensible parameters, designed in such a way as to restrict possible transformations. Specifically, reflections cannot occur and rotations are limited to +-180 degrees. The parameterisation makes it easier to further external restrictions on the transformations (e.g. not allowing rotations, only allowing isotropic scaling, etc.).