Change order of transformations in RestrictedTransformer
Previously, transformations were applied in order translate, then
rotate, then scale. This is appropriate for transforming an input
image to a 'standardised' (centred, derotated, circularised)
format. However, when applying to a 'standardised' image, to
reintroduce transformations, then order needs to be reversed: scale,
then rotate, then translate. (This is vital, since the original method
cannot produce an elongated object at an angle when given a circular
input.) The latter order is now the default, but the original order is
still available by specifying `reverse=True`.