Added option to Affine and Restricted to preserve flux
Previous (and still default) method preserves the value at a given
(mapped) position, but results in a change in the total flux (sum of
all pixel values). In some circumstances (e.g. astronomy) it is more
appropriate to maintain the total flux. The change in flux arises in
two ways: the first is a rescaling in the affine transformation. The
second is if the input is mapped onto an output with a different
size. Both of these are accounted for by setting the new option
`preserve_flux=True`.